renamed images and inserted only one to Cast and Crew
