+++ src/features/MoviePage/Content/Cast/index.js
@@ -1,133 +1,127 @@
 import { Content, Text, Title, ImageTile, Picture, Character, Name, Role } from './styled';
 import { Row, PersonTile } from '../../../../components/Tile';
-import first from '../../images/1.png';
-import firstMobile from '../../images/1m.png';
-import secend from '../../images/2.png';
+import first from '../../images/person.png';
+import secend from '../../images/person2.png';
 import profile from '../../images/profile.png';
-import { useScreenWidth } from '../../../../useScreenWidth';
 
-const Cast = () => {
-    const screenWidth = useScreenWidth();
-
-    return (
-        <Content>
-            <Title>Cast</Title>
-            <Row>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={secend} />
-                        <Character>
-                            <Name>
-                                <Text>Donnie Yen</Text>
-                                <Role>Komandor Tung</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Jason Scott Lee</Text>
-                                <Role>Böri Khan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Jet Li</Text>
-                                <Role>Böri Khan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-            </Row>
-            <Row>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={profile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-                <PersonTile>
-                    <ImageTile>
-                        <Picture src={screenWidth > 767 ? first : firstMobile} />
-                        <Character>
-                            <Name>
-                                <Text>Liu Yifei</Text>
-                                <Role>Mulan</Role>
-                            </Name>
-                        </Character>
-                    </ImageTile>
-                </PersonTile>
-            </Row>
-        </Content>
-    )
-};
+const Cast = () => (
+    <Content>
+        <Title>Cast</Title>
+        <Row>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={secend} />
+                    <Character>
+                        <Name>
+                            <Text>Donnie Yen</Text>
+                            <Role>Komandor Tung</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Jason Scott Lee</Text>
+                            <Role>Böri Khan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Jet Li</Text>
+                            <Role>Böri Khan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+        </Row>
+        <Row>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={profile} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+            <PersonTile>
+                <ImageTile>
+                    <Picture src={first} />
+                    <Character>
+                        <Name>
+                            <Text>Liu Yifei</Text>
+                            <Role>Mulan</Role>
+                        </Name>
+                    </Character>
+                </ImageTile>
+            </PersonTile>
+        </Row>
+    </Content>
+);
 
 export default Cast;
